refactor: module tree now mirrors comp_gc ∘ it_gc ∘ orchestration
Split the vague components/ grab-bag and make the layering explicit (all via
git mv, no logic changes):
- garble/ -> comp_gc/, with the Phase-1 builders (ohe, convert) folded in, so the
  computational GC is one named module symmetric with it_gc.
- pipeline.rs and affine.rs moved to crate root (the streaming orchestration + the
  S_aff driver, the top layer).
- crt.rs + bigint.rs moved into a leaf crt/ dir (pure math, no garbling dep).

Tree now reads: foundation (types/label/hash/system/exec) + crypto -> comp_gc /
it_gc -> pipeline -> affine, with crt/ as leaf math.

147 tests pass; clippy clean; off-aarch64 compiles.
