--- src/components/mod.rs
@@ -1,10 +1,0 @@
-/// Affine maps over a primorial ring via CRT decomposition.
-pub mod affine;
-/// Fixed-width 576-bit unsigned integer for CRT output representation.
-pub mod bigint;
-/// Conversions between binary, word, and ring representations.
-pub mod convert;
-/// CRT parameters and reconstruction.
-pub mod crt;
-/// One-hot encoding and scaling.
-pub mod ohe;
